+++ crowdfunding/users/views.py
@@ -3,7 +3,8 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from .models import CustomUser
-from .serializers import CustomUserSerializer
+from .serializers import CustomUserSerializer, CustomUserDetailSerializer
+
 class CustomUserList(APIView):
     def get(self, request):
           users = CustomUser.objects.all()
@@ -26,21 +27,31 @@ class CustomUserDetail(APIView):
 
     def get(self, request, pk):
         user = self.get_object(pk)
-        serializer = CustomUserSerializer(user)
+        serializer = CustomUserDetailSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
         user = self.get_object(pk)
+        self.check_object_permissions(request, user)
         data = request.data
-        serializer = CustomUserSerializer(
+        serializer = CustomUserDetailSerializer(
             instance=user,
             data=data,
             partial=True
         )
         if serializer.is_valid():
             serializer.save()
+            return Response(
+                serializer.data,
+                status=status.HTTP_201_CREATED
+            )
+        return Response(
+            serializer.errors,
+            status=status.HTTP_400_BAD_REQUEST
+        )
 
     def delete(self, request, pk, format=None):
         event = self.get_object(pk)
+        self.check_object_permissions(request, user)
         event.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
